# Conflicts:
#	.github/CODEOWNERS
#	src/main/java/com/gnagnoohc/travel/auth/controller/AuthApiController.java
#	src/main/java/com/gnagnoohc/travel/auth/controller/AuthController.java
#	src/main/java/com/gnagnoohc/travel/auth/dto/SignUpRequest.java
#	src/main/java/com/gnagnoohc/travel/auth/dto/VerifiedSignupEmail.java
#	src/main/java/com/gnagnoohc/travel/auth/exception/EmailVerificationException.java
#	src/main/java/com/gnagnoohc/travel/auth/exception/SignupException.java
#	src/main/java/com/gnagnoohc/travel/auth/mapper/AuthMapper.java
#	src/main/java/com/gnagnoohc/travel/auth/model/EmailVerification.java
#	src/main/java/com/gnagnoohc/travel/auth/model/Member.java
#	src/main/java/com/gnagnoohc/travel/auth/model/MemberLocalAuth.java
#	src/main/java/com/gnagnoohc/travel/auth/service/AuthService.java
#	src/main/java/com/gnagnoohc/travel/auth/service/EmailVerificationService.java
#	src/main/java/com/gnagnoohc/travel/auth/service/GoogleMailService.java
#	src/main/resources/mapper/auth/AuthMapper.xml
#	src/main/resources/static/js/auth/signup.js
#	src/main/webapp/WEB-INF/views/auth/signup-form.jsp
#	src/main/webapp/WEB-INF/views/auth/signupresult.jsp

+++ src/main/java/com/gnagnoohc/travel/community/dto/ImageDto.java
@@ -9,9 +9,9 @@ import lombok.Data;
 @Data
 @Alias("image")
 public class ImageDto {
-	private Long imageId;             // image_id    BIGINT       PK, AUTO_INCREMENT
-    private Long postId;              // post_id     BIGINT       NOT NULL (FK → post.post_id)
+	private int imageId;              // image_id    INT           PK, AUTO_INCREMENT
+    private int postId;               // post_id     INT           NOT NULL (FK → post.post_id)
     private String imageUrl;          // image_url   VARCHAR(500) NOT NULL
-    private Integer sortOrder;        // sort_order  INT          DEFAULT 0
+    private int sortOrder;            // sort_order  INT          DEFAULT 0
     private Timestamp createdAt;  // created_at  DATETIME     DEFAULT CURRENT_TIMESTAMP
 }
